Fix issue 260
